requirements.txt
discord.py-self
aiohttp
asyncio
requests
flask
